requirements.txt
pytest==8.4.2
requests==2.32.5
allure-pytest==2.15.0
faker==37.12.0
pydantic==2.12.3
python-dotenv==1.2.1
